requirements.txt
BeautifulSoup==3.2.1
Django==1.6.5
arrow==0.4.2
python-dateutil==2.2
pytz==2014.4
requests==2.3.0
six==1.7.3
wsgiref==0.1.2
